feat: add stepper enums for state and style definitions
